+++ packages/models-library/src/models_library/function_services_catalog/services/demo_units.py
@@ -1,10 +1,10 @@
 from ...services import (
-    LATEST_INTEGRATION_VERSION,
     ServiceInput,
     ServiceMetaDataPublished,
     ServiceOutput,
     ServiceType,
 )
+from ...services_constants import LATEST_INTEGRATION_VERSION
 from .._key_labels import FUNCTION_SERVICE_KEY_PREFIX
 from .._utils import OM, PC, FunctionServices, create_fake_thumbnail_url
