+++ packages/models-library/src/models_library/function_services_catalog/services/iter_range.py
@@ -1,11 +1,8 @@
 from collections.abc import Iterator
 
 from ...projects_nodes import OutputID, OutputsDict
-from ...services import (
-    LATEST_INTEGRATION_VERSION,
-    ServiceMetaDataPublished,
-    ServiceType,
-)
+from ...services import ServiceMetaDataPublished, ServiceType
+from ...services_constants import LATEST_INTEGRATION_VERSION
 from .._key_labels import FUNCTION_SERVICE_KEY_PREFIX
 from .._utils import OM, FunctionServices, create_fake_thumbnail_url
