+++ packages/models-library/src/models_library/function_services_catalog/services/nodes_group.py
@@ -1,8 +1,5 @@
-from ...services import (
-    LATEST_INTEGRATION_VERSION,
-    ServiceMetaDataPublished,
-    ServiceType,
-)
+from ...services import ServiceMetaDataPublished, ServiceType
+from ...services_constants import LATEST_INTEGRATION_VERSION
 from .._key_labels import FUNCTION_SERVICE_KEY_PREFIX
 from .._utils import OM, FunctionServices
