+++ packages/models-library/src/models_library/function_services_catalog/services/probes.py
@@ -1,10 +1,7 @@
 from typing import Final
 
-from ...services import (
-    LATEST_INTEGRATION_VERSION,
-    ServiceMetaDataPublished,
-    ServiceType,
-)
+from ...services import ServiceMetaDataPublished, ServiceType
+from ...services_constants import LATEST_INTEGRATION_VERSION
 from .._key_labels import FUNCTION_SERVICE_KEY_PREFIX
 from .._utils import OM, WVG, FunctionServices, create_fake_thumbnail_url
