+++ packages/models-library/src/models_library/services.py
@@ -8,9 +8,9 @@ from .boot_options import BootOption, BootOptions
 from .emails import LowerCaseEmailStr
 from .services_authoring import Author, Badge
 from .services_base import ServiceBase, ServiceKeyVersion
+from .services_constants import ANY_FILETYPE, LATEST_INTEGRATION_VERSION
 from .services_enums import ServiceType
 from .services_io import BaseServiceIOModel, ServiceInput, ServiceOutput
-from .services_regex import ANY_FILETYPE, LATEST_INTEGRATION_VERSION
 from .services_types import (
     DynamicServiceKey,
     RunID,
