+++ packages/models-library/src/models_library/services_constants.py
@@ -1,0 +1,6 @@
+from typing import Final
+
+LATEST_INTEGRATION_VERSION: Final[str] = "1.0.0"
+
+
+ANY_FILETYPE: Final[str] = "data:*/*"
