+++ packages/models-library/src/models_library/services_io.py
@@ -10,7 +10,8 @@ from pydantic import (
     validator,
 )
 
-from .services_regex import ANY_FILETYPE, PROPERTY_TYPE_RE
+from .services_constants import ANY_FILETYPE
+from .services_regex import PROPERTY_TYPE_RE
 from .services_types import FileName, ServicePortKey
 from .services_ui import Widget
 from .utils.json_schema import (
